Replaced old branch naming format with issue tracker
I have replaced the old naming conventions for my branches for this project to now use issue tracker id's for better organization of this project which would enable for better understanding in the future when its contents are reviewed. The current issue # is 13 which pertains to working on the mileston project. More specifically, the impementation of switching between players and restaring the game.