Completed lecture 70 OOP
Issue marks the continuation of this section to document the progress made over time and highlight key points in the section that can be used for later Python development tasks related to object oriented programming. It is the section after the first course project. Section was previously visted months ago with the two first lectures completed at that time.