Made the changes to effectively use the full_board_check() funnction
Was able to make the necessary changes in code to use the full_board_check function
to  produce the deisred reuslts of checking the game board for all X's and O's. Also
impleneted the win_check function in the code to check if there is a winner. win_check
needs to be properly implemented